Refactor review validation
